feat: get do token usando base de dados

+++ Core/IBGEExplorer.Account/Entities/Role.cs
@@ -2,10 +2,10 @@
 
 namespace IBGEExplorer.Account.Entities;
 
-public class Role : Entity<short>
+public class Role : Entity<int>
 {
     public string? Name { get; set; }
-    public short RoleId { get; set; }
+    public int RoleId { get; set; }
 
     public List<UserRole>? UserRoles { get; set; }
 }
