+++ Core/IBGEExplorer.Account/Entities/UserRole.cs
@@ -2,10 +2,10 @@
 
 namespace IBGEExplorer.Account.Entities;
 
-public class UserRole : Entity<short>
+public class UserRole : Entity<int>
 {
     public int UserId { get; set; }
-    public short RoleId { get; set; }
+    public int RoleId { get; set; }
 
     public User? User { get; set; }
     public Role? Role { get; set; }
